STAT-65: More renaming
