+++ docs/usage.md
@@ -1,7 +1,7 @@
 # Usage
 
 ```{eval-rst}
-.. click:: dapla_hurtigstart_cli.__main__:main
-    :prog: hack2022-dapla-hurtigstart
+.. typer:: ssb-project.__main__:main
+    :prog: ssb-project
     :nested: full
 ```
